requirements.txt
PyQt6>=6.5.0
cryptography>=41.0.0
argon2-cffi>=23.1.0
